refactor(web): share organization selector across admin forms
Extract the searchable organization combobox built for the credit-grants
form into a reusable OrganizationCombobox, and use it in the enterprise
contract form in place of the free-text slug input.

- New OrganizationCombobox (Popover + Command) with its own i18n namespace.
- New adminOrganizationService.listOrganizations as the single source for the
  admin org list; credit-grant-admin.service no longer duplicates it.
- Enterprise contract new/edit pages now load organizations and pass them to
  the form; the org field is a picker (still disabled in edit mode), and its
  label changed from "Slug" to "Organization".
